Fixed assignment operator terminology
It's not accurate to call the `assignment operator` `=` and `equal` operator, since it might be confused with the `equality operators`, eg `==`.

For the very same reason changed the `Minus Equals` wording to characters `-=` and titled the challenges with its commonly used name: compound or augment assignment (see https://en.wikipedia.org/wiki/Augmented_assignment)

What do you think of this clarification?

Thanks.