Added link to a book on Algorithms (#22308)
This is the of the programming language-independent text that helped establish computer algorithms as a discipline of computer science. The text incorporates the latest research and state-of-the-art applications, bringing this classic to the forefront of modern computer science education. A major strength of this text is its focus on design techniques rather than on individual algorithms.