Updated index.md to change the grammatical mistakes (#32494)
* Updated index.md to change the grammatical mistakes

In the line no. 6, added an "an" before the word "element". Similar corrections in line no. 8, 9, and 10. In line no. 8, "then" is added in the line to make a proper meaning in completing the sentence. The word "to the" is added before the words such as "top", "left" and "bottom" in line nos. 9 and 10 to make a proper sentence.

* Update index.md

* Added a , before the then statement